Move logic to employment class, as profile_employment will not use these validations
