+++ app/models/employment.rb
@@ -5,6 +5,8 @@ class Employment < EmploymentRecord
 
   # KSIE dictates that we need a reason_for_leaving even for current role
   validates :reason_for_leaving, :main_duties, presence: true, if: -> { job? }
+  validates :main_duties_words, length: { maximum: MAIN_DUTIES_MAX_WORDS }, if: -> { main_duties.present? }
+  validates :reason_for_leaving_words, length: { maximum: REASON_FOR_LEAVING_MAX_WORDS }, if: -> { reason_for_leaving.present? }
 
   def duplicate
     # dup does a shallow copy, but although it "doesn't copy associations" according to the
@@ -12,5 +14,15 @@ class Employment < EmploymentRecord
     dup.tap do |record|
       record.assign_attributes(job_application: nil)
     end
+  end
+
+  private
+
+  def main_duties_words
+    main_duties.strip.split(/\s+/)
+  end
+
+  def reason_for_leaving_words
+    reason_for_leaving.strip.split(/\s+/)
   end
 end
